refactor: lib/firebaseに分離
